feat(tmp buf): Add `use_tmp_buffer` option
To fix filetype even if some script has already called `setfiletype`
somewhere and made calling once more not work for overriding, create
a temporary buffer and run detection for a fixed path on it instead
of same buffer.
Currently do so only if `g:chezmoi#use_tmp_buffer` is set for not
enough tests.