fix(tmp buf): Detect shebang
The `put = getbufline(<bufnr>, 1, '$')` copies lines with preserving
existing lines, so the destination buffer will still have a first empty
line and a shebang (e.g. `#!/bin/sh`) -- that tells which interpreter
of script file -- will be in second line and invalid. Duo to that,
filetype detection misses the shebang and doesn't work correctly. So move
it from the second line to the first by deleting a first empty line.