fix(detection): Ignore a `Quickfix` list
Opening a Quickfix list buffer (e.g. by calling `vimgrep`) shows
an error of E788 if the `use_tmp_buffer` option is On. Vim doesn't
allowed to create a new buffer while it's opening a Quickfix.
Whether or not the error is, it shouldn't need to fix highlighting
for any Quickfix list as couldn't be a source file of `chezmoi`.
So ignore it.