fix(tmp buf): Keep clipboard contents
If Vim synchronizes contents in the un-named register and the
system clipboard via setting the `clipboard` option, making a
temporary buffer breaks the clipboard contents because it
overwrites the un-named register with original file contents.
So keep the contents by storing into a temporary variable and
restoring it after file type detection.